refactor: Follow, Like 테스트코드 패키지 위치 수정
